fix Calling- and Called-Station-Id session attributes
The session state was expected to contain without the -Id suffix.
However, most AAA systems use this attribute with the suffix and
is simply more natural to use that for the sessions state as well.

The only release consumer of this project was as using the -Id
suffix in the session state.