Initial version of Diameter Accounting RFC 7155
For now, authentication is mocked to return success. `Start`, `Interim`
and `Stop` accounting methods are implemented with basics fields from
`ergw_aaa_session`.

Here is simple test for diameter which does 2 things:
1. Checks CER and CEA between ergs-aaa-diameter client and
test-diameter-server
2. Creates `ergw_aaa_sessions` and tries to start accounting for it.

To test that Diameter requests were sent or received service statistics
for ergw-aaa-diameter are used but we don't check response's payload for
now.