rework the AAA application configuration and session logic
We now have handlers that implement a specific a protocol or
referenc point (e.g. RADIUS or DIAMETER NASREQ).
Services that use on of the protocol handler and can modify
the configuration of a handler.
The interface to the consumer is the 'AAA application'. It
initilizes a session by invoking the initialization of
a number of services and has multiple procedures (e.g.
authentication, accounting, ...). Each procedure can have its
own configuration.