Rework applying the attribute_map
Some values resulting from applying the given attribute_map are
charlists, some are binaries, etc. A charlist leads to a crash when
treating it like a binary ( <<Res/binary, Value/binary>> ).

Converting every value to a binary and then appending it to an existing
binary is tedious.

Instead, we use lists:map/2 and iolist() to collect all pieces.