Add mechanism to fake procedure answers and use it to handle OCS downs
Preloading answers in the static handler is to limited to deal with
the different answers needed for different procedures.
Add a mechanism to load per procedure answers in the static handler.
Use the same mecahnism in the Ro handler to define a per procedure
answer in case the OCS is down.