use session event of Multiple-Services-Credit-Control updates
The Multiple-Services-Credit-Control attribute is not static, it
conveys a change of granted, used and requested quotas (in both
direction, to and from the OCS).
Putting it like a static state attrubute into the session makes it
difficult to work with the state change. Instead report MSCCs through
the session event mechanism.

Compare MSCC responses to requests and inject error responses for
every Rating-Group that hasn't been replied to. This is mostly usefull
for staticly configure replies.