disable OTP Diameter built-in re-transmit handling
The OTP Diameter stack tries to re-transmit a request to another
peer if the FSM for the original host transitions from UP to
something else (SUSPECT or DOWN). None the callbacks (peer_down/2,
handle_error or handle_result) are called when that happens.

This cause the request tracking in the load estimation logic to
go out of sync, ultimately breaking load distribution completely.

This change disables the built-in re-transmission all together and
let the application handle it.

Gy host stickiness and fail-over support is still incorrect (as it
was before).