Use accounting trigger for nasreq accounting
Using the same accounting trigger as radius fixes the missing volume
report AVPs : Accounting-Input-Octets and Accounting-Output-Octets
in interim and stop NASREQ ACR.

The minor limitation of this solution is that both NASREQ and RADIUS
accounting with different interim report intervals can not be used the
same time in a deployment. It is a very unlikely case though.