make the session init more asynchronous
The session init phase can take significant time. This would cause
a backlog in the session supervisor when many sessions where started
at once. This becomes visible as a significant grows of the mailbox
of the supervisor process.

Also increase the outstanding requests limit for the rate limit
test. The inproved startup time also means the traffic in that
test will be much more bursty and will run into that limit.