Add Diameter peer rate limit adjustment functions
This peer below has reached 10 of 100. Sometimes a peer gets stuck at 100 of 100.
The idea would be to have something simple that can be called in a single line,
returning a map with
<<"m4sxhpsrm651">> => {peer,10,100,-571873157298,200,5,199}
entries,
and an adjust peer function that can lower this.
The idea would be to make this available to a custom resource at some point.