Avoid taking InOctets and OutOctets into TDV
Existing InOctets and OutOctets session keys should not
be used to build the Traffic-Data-Volumes item in Rf.
In some corner cases involving combined RADIUS and Rf
traffic it may cause a crash in the AAA session.

Tests for both a simple case including these keys
in a request, and a combined RADIUS/Rf request, are
provided.