[ct] change the outstanding_requests setting in the Rf and Gy suite
The setting is used by the rate limit checks. However, those tests are
designed to check for handling the rate limit only. The outstanding
requests limits is not checked and hitting it interferes with the rate
limit check, leading to spurious failure.

Raise the value to a limit that it will no longer affect the rate
limit tests.